refactor: 상수 패키지 생성
- 상수 클래스/열거형을 모아놓은 Constants 패키지를 생성.
- 상수 관련 클래스들을 해당 패키지로 이동. 이에 따른 코드 수정.
