+++ src/main/java/racingcar/constants/Message.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.constants;
 
 public enum Message {
     NAMEINPUT("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
