+++ src/main/java/racingcar/model/Position.java
@@ -1,9 +1,9 @@
 package racingcar.model;
 
-import static racingcar.Constants.SLASH;
-import static racingcar.PositionConstant.FORWARD_THRESHOLD;
-import static racingcar.PositionConstant.VALUE_HIGH_BOUND;
-import static racingcar.PositionConstant.VALUE_LOW_BOUND;
+import static racingcar.constants.Constants.SLASH;
+import static racingcar.constants.PositionConstant.FORWARD_THRESHOLD;
+import static racingcar.constants.PositionConstant.VALUE_HIGH_BOUND;
+import static racingcar.constants.PositionConstant.VALUE_LOW_BOUND;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import racingcar.dto.ForwardState;
